+++ README.md
@@ -50,16 +50,13 @@
     - [x] 무협
     - [x] BL
   - [x] 탭 선택 시, 화면 새로고침이 일어나지 않는다.
-- step3
   - [x] 탭을 누르면 하위 콘텐츠가 변경되어 노출된다
   - [x] 선택 된 탭 색상 변경
-  - [x] 탭을탭 선택 시, 화면 새로고침이 일어나지 않는다.
   - [x] 탭을웹툰 영역을 기본으로 개발하고, 추가 메뉴들도 동작하게 한다.
   - [x] 탭을웹툰의 경우 요일별 탭 선택이 동작해야 한다.
   - [x] 무한 슬라이딩 (Carousel)
   - [x] 자동 슬라이딩 기능
-  - [x] 자동 또는 수동 슬라이딩시 우측하단에 페이지 번호 표시
-
-#### 문제점
-
-#### 고민 하고 있는 것
+  - [x] 자동 또는 수동 슬라이딩시 우측하단에 페이지
+  - [x] 클라이언트 정적인 자원요청(html,js,css,image)에 응답할 수 있는 서버 구성
+- step3
+  - [ ] 구현중인 기능에 대해서 모든 데이터 요청을 서버로부터 JSON형태로 받아서 처리하도록 한다. (일부만 처리)
