+++ express/public/js/components/bl.js
@@ -1,4 +1,4 @@
-import { blData } from '../data/bl-data.js';
+import { blData } from '../data/menu-data.js/bl-data.js';
 import { contents } from '../html-template.js';
 
 export default function addHTML() {
