--- express/public/js/data/slide-data.js
@@ -1,87 +1,0 @@
-// 슬라이드 이미지 데이터
-export const slideData = [
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=bieQhm/hzp2mI9n4A/9aqnCcDnI3LGkjcXFC8n00',
-    title: '지수에 관하여',
-    status: 'New',
-    people: '1.6',
-    message: '선배는, 왜 저한테 잘해 주세요?',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=sU3Ip/hzhOd1K291/NlK8pODmELGCbJ8rJdF3v1',
-    title: 'BJ 대마도사',
-    status: 'UP',
-    people: '26',
-    message: '왜 게임 정보들이 보이는 거지? ',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=AZSyM/hzp2gPCXnk/88NFfy4lG7YzhEziCUZ2jk',
-    title: '인간시장',
-    status: 'UP',
-    people: '21.8',
-    message: '거짓투성이 네 인생, 내가 빼앗겠어',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=camF4n/hzp2jeu9J5/IcHSCJiuWmjedlUeiDMKY1',
-    title: '악녀를 죽여줘',
-    status: '이벤트',
-    people: '64.7',
-    message: '시즌2 오픈 기념 이벤트',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=bqyJkx/hzp2fwmJDP/qx8LbXn8n6DHxfK0B4sMkk',
-    title: '톱스타, 그 자체',
-    status: 'NEW',
-    people: '1.6',
-    message: '다시 톱스타가 될 수 있을까?',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=fK7X9/hzhOgcggkT/AHKP8zDsUABstUU49go5f1',
-    title: '학사귀환',
-    status: '이벤트',
-    people: '50.3',
-    message: '학사 마현, 중원으로 돌아오다!',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=VkUeK/hzp2dSRMoq/kXLSvCGDb4baPuuBnQzZw1',
-    title: '사내 맞선',
-    status: '이벤트',
-    people: '201.8',
-    message: '외전 런칭! 무료UP + 이벤트!',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=cl0NGc/hzacezTSJp/rNn75CBqttbKDY0Cgpierk',
-    title: '내가 죽였다',
-    status: '이벤트',
-    people: '21.1',
-    message: '자살 아니야. 내가 죽였어',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=bqvePz/hzp2h19oMU/EKbeNnPhw39B0hUixK0J80',
-    title: '생사결 : 여래의 복수',
-    status: '이벤트',
-    people: '58.8',
-    message: '완결기념, 한시한판 + 캐시 이벤트!',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=bR3b12/hzhOiBPnrU/Nj39dPNShs7mQxft22H2NK',
-    title: '이차원 용병',
-    status: 'UP',
-    people: '103.1',
-    message: '영혼을 팔아서라도 취직하고 싶으십니까',
-  },
-  {
-    url: 'https://dn-img-page.kakao.com/download/resource?kid=bGq9qL/hzp2jTafgP/QmhYsGrj7xc41ZRb7D9NHk',
-    title: '낙화유슈',
-    status: '이벤트',
-    people: '45.4',
-    message: '시즌2 런칭기념 캐시 이벤트!',
-  },
-  {
-    url: '	https://dn-img-page.kakao.com/download/resource?kid=n34yt/hzp2fiV1br/0S1hm7bgEXkeXzVxWXnSS0',
-    title: '귀환용사의 골목식당',
-    status: 'New',
-    people: '12.9',
-    message: '신작 런칭기념, 캐시 이벤트!',
-  },
-];
